Add const file and move enum BlockType there
