+++ src/components/text-editor/config.ts
@@ -1,27 +1,6 @@
 import Immutable from 'immutable';
 import { DefaultDraftBlockRenderMap } from 'draft-js';
-
-export enum BlockType {
-  /* Заголовки */
-  h1 = 'header-one',
-  h2 = 'header-two',
-  h3 = 'header-three',
-  h4 = 'header-four',
-  h5 = 'header-five',
-  h6 = 'header-six',
-  /* Цитата */
-  blockquote = 'blockquote',
-  /* Блок с кодом */
-  code = 'code-block',
-  /* Список */
-  list = 'unordered-list-item',
-  /* Нумерованный список */
-  orderList = 'ordered-list-item',
-  /* Сноска */
-  cite = 'cite',
-  /* Простой текст */
-  default = 'unstyled',
-}
+import { BlockType } from '../../const';
 
 const CUSTOM_BLOCK_RENDER_MAP = Immutable.Map({
   [BlockType.cite]: {
