+++ src/const.ts
@@ -1,0 +1,21 @@
+export enum BlockType {
+  /* Заголовки */
+  h1 = 'header-one',
+  h2 = 'header-two',
+  h3 = 'header-three',
+  h4 = 'header-four',
+  h5 = 'header-five',
+  h6 = 'header-six',
+  /* Цитата */
+  blockquote = 'blockquote',
+  /* Блок с кодом */
+  code = 'code-block',
+  /* Список */
+  list = 'unordered-list-item',
+  /* Нумерованный список */
+  orderList = 'ordered-list-item',
+  /* Сноска */
+  cite = 'cite',
+  /* Простой текст */
+  default = 'unstyled',
+}
